+++ RELEASENOTES.md
@@ -10,7 +10,12 @@ Note, the changes I make won't go into effect until the next module release date
 
 ## Changes
 
-### 24.10.1 (not yet implemented)
+### 24.11.15 (not yet implemented)
+ - Several minor changes over the past month, including a bug fix - see issues closed over the past month
+ - Completely revamped the index lookup in the DEV GUI - Modified to use a generic Index Map for all Build Versions.  This will greatly simplify the method no longer requiring me to build a new Index Map for each build.
+  - As part of this, built a new Index Map JSON/XML files in the Catalog folders. Will eventually go back and cleanup the no longer used Index files 
+
+### 24.10.1 (Released to Gallary)
  - Updates for Windows 11 24H2 x64 and ARM64
   - Updated Catalog JSON & XML files
   - Updated to OSD.json file to add options for 24H2